доделал кроме дребезга
